split out overrides functions
